Remy's SNMgr plus I added a log centrl user method
When it is changed on the frontend, the backend gets notified to put it in the log
Tested, works, but is very verbose if you change users a lot.
I could see a future "Clean" function that only save the last "s user" to the file. Not for this project.